Modified javascripts for new public map
