+++ lib/assets/javascripts/cartodb/public_map/public_map_window.js
@@ -1,85 +1,91 @@
+var $ = require('jquery');
+var cdb = require('cartodb.js');
+var PublicMap = require('./public_map_view');
   
-  /** 
-   *  Public map window "view"
-   *
-   */
+/** 
+ *  Public map window "view"
+ *
+ */
 
-  cdb.open.PublicMapWindow = cdb.core.View.extend({
+module.exports = cdb.core.View.extend({
 
-    initialize: function() {
-      this.$body = $(this.el.document.body);
-      this._setupMapDimensions();
-      this._initBinds();
-      this._initViews();
-    },
+  initialize: function() {
+    this.$body = $(this.el.document.body);
+    this.$map = this.$body.find('#map');
+    this._setupMapDimensions();
+    this._initBinds();
+    this._initViews();
+  },
 
-    _initViews: function() {
-      // Map view
-      var opts = _.defaults({ el: this.el.document.body }, this.options);
-      this.map_view = new cdb.open.MapPublic(opts);
-      // this.map_view.bind('map_loaded', this._onMapLoaded, this);
-      this.addView(this.map_view);
-    },
+  _initViews: function() {
+    // Map view
+    this.map_view = new PublicMap(_.defaults({ el: this.$map }, this.options));
+    // this.map_vis.bind('map_error', this._showNotSupportedDialog, this);
+    // this.map_vis.bind('map_loaded', function(vis) { this.trigger('map_loaded', vis, this) }, this);
+    // this.addView(this.map_vis);
+    // this.map_view.bind('map_loaded', this._onMapLoaded, this);
+    this.addView(this.map_view);
+  },
 
-    _initBinds: function() {
-      _.bindAll(this, '_onWindowResize', '_onOrientationChange');
+  _initBinds: function() {
+    _.bindAll(this, '_onWindowResize', '_onOrientationChange');
 
-      this.$el.on('resize', this._onWindowResize);
+    this.$el.on('resize', this._onWindowResize);
 
-      if (!this.el.addEventListener) {
-        this.el.attachEvent('orientationchange', this._onOrientationChange, this);
-      } else {
-        this.el.addEventListener('orientationchange', this._onOrientationChange);
-      }
-    },
+    if (!this.el.addEventListener) {
+      this.el.attachEvent('orientationchange', this._onOrientationChange, this);
+    } else {
+      this.el.addEventListener('orientationchange', this._onOrientationChange);
+    }
+  },
 
-    _onWindowResize: function() {
-      // Resize window
-      this._setupMapDimensions();
-      // Close dialogs
-      cdb.god.trigger("closeDialogs");
-    },
+  _onWindowResize: function() {
+    // Resize window
+    this._setupMapDimensions();
+    // Close dialogs
+    cdb.god.trigger("closeDialogs");
+  },
 
-    _onOrientationChange: function() {
-      // Reset disqus
-      DISQUS.reset({ reload: true });
-      // Resize window orientation
-      this._setupMapDimensions(true);
-    },
+  _onOrientationChange: function() {
+    // Reset disqus
+    DISQUS && DISQUS.reset({ reload: true });
+    // Resize window orientation
+    this._setupMapDimensions(true);
+  },
 
-    // When window is resized, make some changes ;)
-    _setupMapDimensions: function(animated) {
-      var windowHeight = this.$el.height();
-      var h, height, top;
+  // When window is resized, make some changes ;)
+  _setupMapDimensions: function(animated) {
+    var windowHeight = this.$el.height();
+    var h, height, top;
 
-      if (this.options.isMobileDevice) {
+    if (this.options.isMobileDevice) {
 
-        if (windowHeight > 670) {
-          h = 180;
-        } else { // iPhone, etc.
-          h = 100;
-        }
-
-      } else {
-        h = 260;
-      }
-
-      height = windowHeight - h;
-      top    = windowHeight - (h - 67);
-
-      if (animated) {
-        this.$body.find("#map").animate({ height: height }, { easing: "easeInQuad", duration: 150 }); 
-      } else {
-
-        if (this.options.isMobileDevice) {
-          this.$body.find("#map").css({ height: height, opacity: 1 }); 
-        } else {
-          // On non mobile devices
-          this.$body.find("#map").css({ height: windowHeight - 194, opacity: 1 })
-        }
+      if (windowHeight > 670) {
+        h = 180;
+      } else { // iPhone, etc.
+        h = 100;
       }
 
-      if (this.map_view) this.map_view.invalidateMap();
+    } else {
+      h = 260;
     }
 
-  });
+    height = windowHeight - h;
+    top    = windowHeight - (h - 80);
+
+    if (animated) {
+      this.$map.animate({ height: height }, { easing: "easeInQuad", duration: 150 }); 
+    } else {
+
+      if (this.options.isMobileDevice) {
+        this.$map.css({ height: height, opacity: 1 }); 
+      } else {
+        // On non mobile devices
+        this.$map.css({ height: windowHeight - 194, opacity: 1 })
+      }
+    }
+
+    if (this.map_view) this.map_view.invalidateMap();
+  }
+
+});
